Issue 2913811 - fix formatting in doc.
Refactored CSVWriter and now have 100% test coverage of writer and refactored classes.

+++ examples/MockResultSet.java
@@ -1,25 +1,9 @@
-
 import java.io.InputStream;
 import java.io.Reader;
 import java.math.BigDecimal;
 import java.net.MalformedURLException;
 import java.net.URL;
-import java.sql.Array;
-import java.sql.Blob;
-import java.sql.Clob;
-import java.sql.Date;
-import java.sql.NClob;
-import java.sql.Ref;
-import java.sql.ResultSet;
-import java.sql.ResultSetMetaData;
-import java.sql.RowId;
-import java.sql.SQLException;
-import java.sql.SQLWarning;
-import java.sql.SQLXML;
-import java.sql.Statement;
-import java.sql.Time;
-import java.sql.Timestamp;
-import java.sql.Types;
+import java.sql.*;
 import java.util.ArrayList;
 import java.util.Calendar;
 import java.util.Map;
@@ -192,7 +176,7 @@ public class MockResultSet implements ResultSet
 
 	public int getInt(String columnName) throws SQLException
 	{
-		return (int) getCurrentRow();
+		return getCurrentRow();
 	}
 
 	public long getLong(String columnName) throws SQLException
